Oracle classic winrm

+++ builder/oracle/classic/step_add_keys.go
@@ -19,6 +19,11 @@ func (s *stepAddKeysToAPI) Run(_ context.Context, state multistep.StateBag) mult
 	config := state.Get("config").(*Config)
 	client := state.Get("client").(*compute.ComputeClient)
 
+	if config.Comm.Type != "ssh" {
+		ui.Say("Not using SSH communicator; skip generating SSH keys...")
+		return multistep.ActionContinue
+	}
+
 	// grab packer-generated key from statebag context.
 	sshPublicKey := strings.TrimSpace(state.Get("publicKey").(string))
 
@@ -49,10 +54,14 @@ func (s *stepAddKeysToAPI) Run(_ context.Context, state multistep.StateBag) mult
 
 func (s *stepAddKeysToAPI) Cleanup(state multistep.StateBag) {
 	// Delete the keys we created during this run
-	keyName := state.Get("key_name").(string)
+	keyName, ok := state.GetOk("key_name")
+	if !ok {
+		// No keys were generated; none need to be cleaned up.
+		return
+	}
 	ui := state.Get("ui").(packer.Ui)
 	ui.Say("Deleting SSH keys...")
-	deleteInput := compute.DeleteSSHKeyInput{Name: keyName}
+	deleteInput := compute.DeleteSSHKeyInput{Name: keyName.(string)}
 	client := state.Get("client").(*compute.ComputeClient)
 	deleteClient := client.SSHKeys()
 	err := deleteClient.DeleteSSHKey(&deleteInput)
